Desktop/Vaishu/ISB/Capstone project/Submission files/Streamlit/Streamlit/requirements.txt
openai
PyPDF2
spire.PDF
pytesseract
img2table
python-docx
torchvision
transformers
easyocr
streamlit
mysql
mysql-connector-python
sqlalchemy-utils
matplotlib